# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Buddy App" ,
    "description": "Buddy Swagger" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "localhost:3000" ,
  "tags": [
    {
      "name": "User" ,
      "description": "API for Users of the system"
    }
  ] ,
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths": {
    "/user/fileUpload":{
      "post": {
        "tags": [
          "User"
        ],
        "summary": "File upload ",
        "description": "Testing file upload in swagger",
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "file to be uploaded",
            "type": "file"
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "responses" :{
          "201": {
            "description": "File uploaded successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "New user registration",
        "description": "Registering a new user to the system",
        "parameters" :[
          {
            "name": "user data",
            "in": "body",
            "description": "body of user registration data",
            "schema": {
              "$ref": "#/definitions/UserRegistration"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "schema": {
              "$ref": "#/responses/userRegistration"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User login route",
        "description": "Login route for user",
        "parameters": [
          {
            "name": "login data",
            "in": "body",
            "description": "Login information",
            "schema" :{
              "$ref": "#/definitions/userLogin"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description" :"User logged in successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "UserRegistration": {
      "required": [
        "email",
        "password",
        "role",
        "name",
        "address",
        "mobile_no"
      ],
      "properties": {
        "email": {
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "mobile_no": {
          "type": "string"
        }
      }
    },
    "userLogin": {
      "required": [
        "email",
        "password"
      ],
      "properties" :{
        "email":{
          "type": "string"
        },
        "password":{
          "type": "string"
        }
      }
    }
  },
  "responses": {
    "userRegistration": {
      "type": "object",
      "properties":{
        "success": {
          "type": "integer"
        },
        "response": {
          "type": "string"
        }
      }
    }
  }
}